VerticalBarChart added
