+++ src/components/VerticalBarChart/VerticalBarChart.scss
@@ -1,4 +1,4 @@
-.NewRelicErrorBreakdown {
+.VerticalBarChart {
   overflow: visible;
   .recharts-tooltip-wrapper {
     z-index: 10;
